requirements.txt
python-telegram-bot==8.0
click
tinydb==3.5.0
schedule==0.4.3
psutil